Added support for PHP 8.4 and PHP 8.5 (#618)
* Add support for PHP 8.4 and 8.5 in composer.json; extend testing matrix in php.yml

* added dev verison of apigee edge

* added github tojen for github action test

* reverted github action code from github test file

* added new dev branch of client php

* version mismatch fix

* Update apigee-client-php version constraints

* fixed the Deprecated function str_getcsv() function issue

* resolved the php8.5 deprecation issue

* resolved deprecation null issue

* fix: capture DateTimeImmutable::setTimezone() return value in tests

* fixed curl error issue

* fix the curl error stale element reference issue

* fixed sting type issue

* fixed deprecation null issue of strtolower

* resolved merge conflict

* Added Updated Version of apigee edge, client php and mockclient php version

* reverted chromedriver version in php yml file

* Updated New Apigee Edge Version

+++ modules/apigee_m10n_add_credit/src/EventSubscriber/CommerceOrderTransitionSubscriber.php
@@ -101,7 +101,7 @@ class CommerceOrderTransitionSubscriber implements EventSubscriberInterface {
     $totals = $this->getCreditTotalsForOrder($order);
 
     foreach ($totals as ['target' => $target, 'amount' => $amount]) {
-      if (!empty((double) $amount->getNumber())) {
+      if (!empty((float) $amount->getNumber())) {
         if (\Drupal::service('apigee_m10n.monetization')->isOrganizationApigeeXorHybrid()) {
           // Use a custom adjustment type because it can support a credit or a debit.
           $job = new BalanceAdjustmentJobX($target, new Adjustment([
