+++ modules/apigee_m10n_add_credit/tests/src/Functional/AddCreditFunctionalTestBase.php
@@ -158,7 +158,7 @@ class AddCreditFunctionalTestBase extends MonetizationFunctionalTestBase {
     $new_balance = \Drupal::service('apigee_m10n.sdk_controller_factory')
       ->developerBalanceController($developer)->getByCurrency('USD');
 
-    static::assertSame((double) $amount, $new_balance->getAmount());
+    static::assertSame((float) $amount, $new_balance->getAmount());
   }
 
 }
